Warnings removed + alphabet restructuration.

+++ src/Bpp/Phyl/Mapping/LaplaceSubstitutionCount.h
@@ -68,7 +68,7 @@ namespace bpp
     mutable RowMatrix<double> m_;
 	
   public:
-    LaplaceSubstitutionCount(const SubstitutionModel* model, int cutOff) :
+    LaplaceSubstitutionCount(const SubstitutionModel* model, size_t cutOff) :
       AbstractSubstitutionCount(new TotalSubstitutionRegister(model->getAlphabet())),
       model_        (model),
       cutOff_       (cutOff),
